Move the redact processor into its own x-pack plugin
